backend/requirements.txt
annotated-types==0.7.0
asgiref==3.11.0
attrs==25.4.0
autobahn==25.12.2
Automat==25.4.16
blis==0.7.11
catalogue==2.0.10
cbor2==5.8.0
certifi==2026.1.4
cffi==2.0.0
channels==4.3.2
channels_redis==4.3.0
charset-normalizer==3.4.4
click==8.3.1
cloudpathlib==0.16.0
confection==0.1.5
constantly==23.10.4
cryptography==46.0.3
cymem==2.0.13
daphne==4.2.1
defusedxml==0.7.1
dj-rest-auth==5.0.2
Django==4.2.7
django-allauth==0.58.2
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
Faker==40.1.2
gunicorn==25.0.3
hyperlink==21.0.0
idna==3.11
Incremental==24.11.0
Jinja2==3.1.6
joblib==1.5.3
langcodes==3.5.1
MarkupSafe==3.0.3
msgpack==1.1.2
murmurhash==1.0.15
numpy==1.26.4
oauthlib==3.3.1
packaging==26.0
pandas==2.1.3
pdfminer.six==20221105
pdfplumber==0.10.3
Pillow==10.1.0
preshed==3.0.12
psycopg2-binary==2.9.9
py-ubjson==0.16.1
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.10.1
pyOpenSSL==25.3.0
PyPDF2==3.0.1
pypdfium2==5.3.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2025.2
redis==7.1.0
requests==2.32.5
requests-oauthlib==2.0.0
scikit-learn==1.8.0
scipy==1.17.0
service-identity==24.2.0
setuptools==80.10.2
six==1.17.0
smart-open==6.4.0
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlparse==0.5.5
srsly==2.5.2
thinc==8.2.5
threadpoolctl==3.6.0
tqdm==4.67.1
Twisted==25.5.0
txaio==25.12.2
typer==0.9.4
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
ujson==5.11.0
urllib3==2.6.3
wasabi==1.1.3
weasel==0.3.4
whitenoise==6.11.0
zope.interface==8.2
